[backend] MVC 패턴을 적용하여 코드 구조화
