+++ backend/app/models/models.py
@@ -1,0 +1,70 @@
+from datetime import datetime, timezone
+from sqlalchemy import Column, Integer, String, ForeignKey, DateTime, Text, UniqueConstraint
+from sqlalchemy.orm import relationship
+from ..database.database import Base
+
+class User(Base):
+    __tablename__ = "users"
+    id = Column(Integer, primary_key=True, index=True)
+    username = Column(String, unique=True, index=True)
+    hashed_password = Column(String)
+    profile_image = Column(String, nullable=True)
+    created_at = Column(DateTime, default=lambda: datetime.now(timezone.utc))
+    
+    # Relationships
+    posts = relationship("Post", back_populates="user")
+    comments = relationship("Comment", back_populates="user")
+
+class Post(Base):
+    __tablename__ = "posts"
+    id = Column(Integer, primary_key=True, index=True)
+    content = Column(Text, nullable=False)
+    owner_id = Column(Integer, ForeignKey("users.id"))
+    likes = Column(Integer, default=0)
+    created_at = Column(DateTime, default=lambda: datetime.now(timezone.utc))
+    updated_at = Column(DateTime, default=lambda: datetime.now(timezone.utc), onupdate=lambda: datetime.now(timezone.utc))
+    
+    # Relationships
+    user = relationship("User", back_populates="posts")
+    comments = relationship("Comment", back_populates="post", cascade="all, delete-orphan")
+
+class Comment(Base):
+    __tablename__ = "comments"
+    id = Column(Integer, primary_key=True, index=True)
+    content = Column(Text, nullable=False)
+    post_id = Column(Integer, ForeignKey("posts.id", ondelete="CASCADE"))
+    owner_id = Column(Integer, ForeignKey("users.id"))
+    created_at = Column(DateTime, default=lambda: datetime.now(timezone.utc))
+    updated_at = Column(DateTime, default=lambda: datetime.now(timezone.utc), onupdate=lambda: datetime.now(timezone.utc))
+    
+    # Relationships
+    post = relationship("Post", back_populates="comments")
+    user = relationship("User", back_populates="comments")
+
+class Follower(Base):
+    __tablename__ = "followers"
+    id = Column(Integer, primary_key=True, index=True)
+    follower_id = Column(Integer, ForeignKey("users.id"))
+    following_id = Column(Integer, ForeignKey("users.id"))
+    created_at = Column(DateTime, default=lambda: datetime.now(timezone.utc))
+    
+    # Add a unique constraint to prevent duplicate follows
+    __table_args__ = (
+        UniqueConstraint('follower_id', 'following_id', name='_follower_following_uc'),
+    )
+
+class Like(Base):
+    __tablename__ = "likes"
+    id = Column(Integer, primary_key=True, index=True)
+    user_id = Column(Integer, ForeignKey("users.id"))
+    post_id = Column(Integer, ForeignKey("posts.id", ondelete="CASCADE"))
+    created_at = Column(DateTime, default=lambda: datetime.now(timezone.utc))
+    
+    # Add a unique constraint to prevent duplicate likes
+    __table_args__ = (
+        UniqueConstraint('user_id', 'post_id', name='_user_post_like_uc'),
+    )
+    
+    # Relationships
+    user = relationship("User")
+    post = relationship("Post")
