+++ backend/app/routers/auth.py
@@ -1,0 +1,42 @@
+from fastapi import APIRouter, Depends, HTTPException
+from sqlalchemy.orm import Session
+from datetime import timedelta
+
+from ..database.database import get_db
+from ..models.models import User
+from ..models.schemas import UserCreate, UserLogin, Token
+from ..security.auth import get_password_hash, verify_password, create_access_token, create_refresh_token
+
+router = APIRouter(tags=["Authentication"])
+
+@router.post("/signup", response_model=dict)
+def signup(user: UserCreate, db: Session = Depends(get_db)):
+    # Check if user already exists
+    existing_user = db.query(User).filter(User.username == user.username).first()
+    if existing_user:
+        raise HTTPException(status_code=400, detail="Username already registered")
+
+    # Hash password and save user
+    hashed_password = get_password_hash(user.password)
+    new_user = User(username=user.username, hashed_password=hashed_password)
+    db.add(new_user)
+    db.commit()
+    db.refresh(new_user)
+
+    return {"message": "User created successfully"}
+
+@router.post("/login", response_model=Token)
+def login(user: UserLogin, db: Session = Depends(get_db)):
+    # Retrieve user from database
+    existing_user = db.query(User).filter(User.username == user.username).first()
+    if not existing_user or not verify_password(user.password, existing_user.hashed_password):
+        raise HTTPException(status_code=401, detail="Invalid username or password")
+
+    # Generate tokens
+    access_token_expires = timedelta(minutes=30)
+    access_token = create_access_token(
+        data={"sub": existing_user.id}, expires_delta=access_token_expires
+    )
+    refresh_token = create_refresh_token(data={"sub": existing_user.id})
+
+    return {"access_token": access_token, "refresh_token": refresh_token, "token_type": "bearer"}
